python_alfresco_api/raw_clients/alfresco_workflow_client/pyproject.toml
[tool.poetry]
name = "workflow_client"
version = "1"
description = "A client library for accessing Alfresco Content Services REST API"
authors = []
readme = "README.md"
packages = [
    {include = "workflow_client"},
]
include = ["CHANGELOG.md", "workflow_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.9"
httpx = ">=0.23.0,<0.29.0"
attrs = ">=22.2.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
